process text file
for testing purposes, we can use this text file.
you can add as many processes and as many cpu bursts as needed.
BUT must be in the correct format. make sure TNCPU indicates actual
number of total CPU bursts, the process finishes with a CPU burst, and
no white spaces at the end of the text file :)
fixes issue #1